### 2025-06-10 #710 @phuongquan
mantis - Multiple Time Series Scanner
Submitting Author Name: T. Phuong Quan
Submitting Author Github Handle: @phuongquan
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/phuongquan/mantis
Version submitted: 0.3.0
Submission type: Standard

Version accepted: TBD

---

```
Package: mantis
Title: Multiple Time Series Scanner
Version: 0.3.0
    person(c("T.", "Phuong"), "Quan", email = "phuong.quan@ndm.ox.ac.uk",
        role = c("aut", "cre"), comment = c(ORCID = "0000-0001-8566-1817")),
    person(family = "University of Oxford", role = "cph"),
    person(family = "National Institute for Health Research (NIHR)", role = "fnd")
    )
Description: Generate interactive html reports that enable quick visual review of multiple
    related time series stored in a data frame. For static datasets, this can help to identify any temporal
    artefacts that may affect the validity of subsequent analyses. For live data feeds,
    regularly scheduled reports can help to pro-actively identify data feed problems
    or unexpected trends that may require action. The reports are self-contained and shareable
    without a web server.
    https://phuongquan.github.io/mantis/
    reactable,
    dygraphs,
    xts,
    scales,
    htmltools
    covr,
```

It helps to identify any temporal artefacts or unexpected trends in collections of related time series. It does this by generating html reports that enable quick visual review of the data. This is important for ensuring the validity of any subsequent analyses based on the data.

There are two main target audiences:

1) researchers who analyse data from large, temporal datasets, particularly routinely-collected data such as electronic health records. The package helps them to quickly check for temporal biases in their data before embarking on their main analyses, increasing the quality of their studies as well as trust in the scientific process.

2) managers of live data feeds that are used as a data source for downstream analyses. Regular inspection of data that is frequently updated will help to identify any issues early and enable a timely response to rectify issues such as missing data. While a set of validation checks in a data pipeline could also work in this circumstance, the benefit of this package is the ability to check for temporal anomalies that are obvious to the human eye but that are difficult to capture programmatically.

I previously created a related package [daiquiri](https://github.com/ropensci/daiquiri) which was accepted at ROpenSci in 2022, which creates data quality reports for temporal, record-level datasets. This new package was developed as a response to users who wanted to visually review datasets that were essentially a collection of time series in a data frame, as opposed to record-level, non-numeric data that requires aggregation, (which is what `daiquiri` is designed for).

There are other packages that do similar things, such as visually inspecting data frames ([visdat](https://github.com/ropensci/visdat)) or running validation checks on them ([assertr](https://github.com/tonyfischetti/assertr), [pointblank](https://github.com/rstudio/pointblank)), but to my knowledge there are none which assist in identifying anomalous _temporal_ changes.

Yes

https://github.com/ropensci/software-review/issues/693

All passed
