+++ pkgdown.yml
@@ -8,4 +8,4 @@ articles:
   articles/example-report-static-heatmap: example-report-static-heatmap.html
   articles/example-report-static-multipanel: example-report-static-multipanel.html
   mantis: mantis.html
-last_built: 2025-06-10T13:57Z
+last_built: 2025-06-10T15:00Z
