### 2025-06-10 cad1587
tidy readme
