### 2025-06-10 30c2640
tidy readme
