### 2025-06-25 #710 @beatrizmilz
- [?] **Tests**: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)?
---

- Installation worked with no errors here.

- About the testing, In the [devguide](https://devguide.ropensci.org/pkg_building.html#testing), is said:
> For testing your functions creating plots, we suggest using [vdiffr](https://vdiffr.r-lib.org/), an extension of the testthat package that relies on [testthat snapshot tests](https://testthat.r-lib.org/articles/snapshotting.html).

But in this package, from what I have experimented, all the visualization are interactive and generated inside an html report output. These are the tests that I found for the function that creates the visualization: https://github.com/phuongquan/mantis/blob/main/tests/testthat/test-plots-interactive.R

So I'm not sure what is the best practise in this case! I will try to find out about this.

- Question: I have experimented some functions, such as `mantis_report()`. I got an `.html` output, generated from a `.Rmd`, but the `.Rmd` was not stored in the project. Is that the intention?

- The contributing information is basic, but after the package is accepted you can use [the template from rOpenSci dev guide](https://devguide.ropensci.org/maintenance_collaboration.html). Anyway, I recommend reading the page about ["Collaboration Guide" in the dev guide](https://devguide.ropensci.org/maintenance_collaboration.html).

- While I was using the package for this, I got an error (using code from the website). It would be good to check, I'm using v0.2.0:

``` r
library(mantis)

data("example_prescription_numbers")

inspec_tabbed2 <- inputspec(
  timepoint_col = "PrescriptionDate",
  item_cols = c("Antibiotic", "Spectrum", "Location"),
  value_col = "NumberOfPrescriptions",
  tab_col = "Location",
  timepoint_unit = "day"
)
#> Error in inputspec(timepoint_col = "PrescriptionDate", item_cols = c("Antibiotic", : unused argument (timepoint_unit = "day")
```

---
