### 2025-06-26 e4e02b1
update latest release version in readme
