### 2025-06-26 3b8e129
add tests for static plots
