### 2025-06-27 #710 @phuongquan
Hi @beatrizmilz!

Thanks for taking this up as handling editor. In response to your comments:

> About the testing, In the [devguide](https://devguide.ropensci.org/pkg_building.html#testing), is said:
>
> For testing your functions creating plots, we suggest using [vdiffr](https://vdiffr.r-lib.org/), an extension of the testthat package that relies on [testthat snapshot tests](https://testthat.r-lib.org/articles/snapshotting.html).
>
> But in this package, from what I have experimented, all the visualization are interactive and generated inside an html report output. These are the tests that I found for the function that creates the visualization: https://github.com/phuongquan/mantis/blob/main/tests/testthat/test-plots-interactive.R
>

There are two static visualisations as well as the (default) interactive one. I have added some new tests in a new test-plots-static.R file which uses vdiffr to test the static plots. Regarding the interactive visualisation, please let me know if you find out how these can be reliably tested.

> Question: I have experimented some functions, such as mantis_report(). I got an .html output, generated from a .Rmd, but the .Rmd was not stored in the project. Is that the intention?

Yes, for a standard report generated using mantis_report(), the intention is just to create the .html output, using the .Rmd stored in the mantis inst/rmd folder. This .Rmd should not be handled directly by the user.

For bespoke reports, the user is expected to create their own .Rmd, and insert bespoke_rmd_xx functions into chunks as required.

> The contributing information is basic, but after the package is accepted you can use [the template from rOpenSci dev guide](https://devguide.ropensci.org/maintenance_collaboration.html). Anyway, I recommend reading the page about ["Collaboration Guide" in the dev guide](https://devguide.ropensci.org/maintenance_collaboration.html).

Thanks, I can add more detail to this when appropriate.

> While I was using the package for this, I got an error (using code from the website). It would be good to check, I'm using v0.2.0:
```
library(mantis)

data("example_prescription_numbers")

inspec_tabbed2 <- inputspec(
  timepoint_col = "PrescriptionDate",
  item_cols = c("Antibiotic", "Spectrum", "Location"),
  value_col = "NumberOfPrescriptions",
  tab_col = "Location",
  timepoint_unit = "day"
)
#> Error in inputspec(timepoint_col = "PrescriptionDate", item_cols = c("Antibiotic", : unused argument (timepoint_unit = "day")

```
Ah, I forgot to update the readme and github release after upping the version to v0.3.0 (which included changing that parameter name). I have updated these now, and if you install v0.3.0 you should no longer get the error.

Let me know if I've missing/misunderstood anything.

Thanks again,
Phuong
