### 2025-06-27 #710 @beatrizmilz
Hi Phuong! I installed the version v0.3.0 and now the error is gone. Great, thanks!

Thanks for for clarifying. I will start inviting reviewers.
