### 2025-06-27 #710 @beatrizmilz
@mpadge Is is OK if I remove the "holding" label?
