### 2025-07-01 d27913c
add rOpenSci badge
