### 2025-07-16 9c9a4d1
Lightweight initial CRAN submission
