### 2025-07-22 dca5f97
remove line breaks from Description (CRAN policy)
