### 2025-07-23 0f3bf83
stop mantis_report() writing to wd by default (CRAN policy). fix #24
