+++ R/report.R
@@ -7,6 +7,12 @@
 #'
 #' @param df A data frame containing multiple time series in long format. See
 #'   Details.
+#' @param file String specifying the desired file name (and path) to save the
+#'   report to. The file name should include the extension ".html". If only a
+#'   file name is supplied, the report will be saved in the current working
+#'   directory. If a path is supplied, the directory should already exist. Any
+#'   existing file of the same name will be overwritten unless
+#'   `add_timestamp` is set to `TRUE`.
 #' @param inputspec [`inputspec()`] object specifying which columns in the
 #'   supplied `df` represent the "timepoint", "item", "value"  and (optionally)
 #'   "tab" for the time series. If a "tab" column is specified, a separate tab
@@ -14,19 +20,17 @@
 #' @param outputspec `outputspec` object specifying the desired format of the
 #'   html table(s). If not supplied, default values will be used.
 #' @param alertspec [`alertspec()`] object specifying conditions to test and
-#'   display
+#'   display.
 #' @param report_title Title to appear on the report.
 #' @param dataset_description Short description of the dataset being shown. This
 #'   will appear on the report.
-#' @param save_directory String specifying directory in which to save the
-#'   report. Default is current directory.
-#' @param save_filename String specifying filename for the report, excluding any
-#'   file extension. If no filename is supplied, one will be automatically
-#'   generated with the format `mantis_report_YYMMDD_HHMMSS`.
+#' @param add_timestamp Append a timestamp to the end of the filename with
+#'   format `_YYMMDD_HHMMSS`. This can be used to keep multiple versions of the
+#'   same report. Default = `FALSE`.
 #' @param show_progress Print progress to console. Default = `TRUE`.
 #' @param ... Further parameters to be passed to `rmarkdown::render()`. Cannot
 #'   include any of `input`, `output_dir`, `output_file`, `params`, `quiet`.
-#' @return A string containing the name and path of the saved report.
+#' @return A string containing the name and full path of the saved report.
 #'
 #' @details The supplied data frame should contain multiple time series in long
 #'   format, i.e.:
@@ -44,10 +48,11 @@
 #'   The `inputspec` parameter maps the data frame columns to the above.
 #' @examples
 #' \donttest{
-#' # create an interactive report in the current directory,
+#' # create an interactive report in the temp directory,
 #' # with one tab per Location
 #' filename <- mantis_report(
 #'   df = example_prescription_numbers,
+#'   file = file.path(tempdir(), "example_prescription_numbers_interactive.html"),
 #'   inputspec = inputspec(
 #'     timepoint_col = "PrescriptionDate",
 #'     item_cols = c("Location", "Antibiotic", "Spectrum"),
@@ -58,17 +63,18 @@
 #'   outputspec = outputspec_interactive(),
 #'   report_title = "Daily antibiotic prescribing",
 #'   dataset_description = "Antibiotic prescriptions by site",
-#'   save_directory = ".",
-#'   save_filename = "example_prescription_numbers_interactive",
 #'   show_progress = TRUE
 #' )
-#' \dontshow{file.remove("./example_prescription_numbers_interactive.html")}
+#'
+#' filename
+#' \dontshow{unlink(filename)}
 #' }
 #'
 #' \donttest{
-#' # create an interactive report in the current directory, with alerting rules
+#' # create an interactive report in the temp directory, with alerting rules
 #' filename <- mantis_report(
 #'   df = example_prescription_numbers,
[…]
