+++ vignettes/articles/example-report-static-heatmap.Rmd
@@ -24,6 +24,7 @@ if (
   filename <-
     mantis::mantis_report(
       df = example_prescription_numbers,
+      file = "../../docs/articles/example-prescriptions-static-heatmap",
       inputspec = mantis::inputspec(
         timepoint_col = "PrescriptionDate",
         item_cols = c("Antibiotic", "Location"),
@@ -33,8 +34,6 @@ if (
       outputspec = mantis::outputspec_static_heatmap(),
       report_title = "Example mantis report - heatmap",
       dataset_description = "Antibiotic prescriptions by site",
-      save_directory = "../../docs/articles",
-      save_filename = "example-prescriptions-static-heatmap",
       show_progress = FALSE
     )
 }
