+++ README.Rmd
@@ -94,9 +94,10 @@ head(example_prescription_numbers)
 ```
 
 ```{r, eval = FALSE}
-# create a report in the current directory, with one tab per Location
+# create a report in the working directory, with one tab per Location
 mantis_report(
   df = example_prescription_numbers,
+  file = "example_prescription_numbers_report.html",
   inputspec = inputspec(
     timepoint_col = "PrescriptionDate",
     item_cols = c("Location", "Antibiotic", "Spectrum"),

[…]
