### 2025-07-24 b511446
Use subdirectory of tempdir for rendering intermediate files instead of output folder. (#24)
