+++ devtesting/test_mantis_report.Rmd
@@ -1,0 +1,34 @@
+---
+title: "test_mantis_report"
+output: html_document
+---
+
+```{r, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+```{r setup}
+# IMPORTANT: First need to Build>Install Package
+
+library(mantis)
+```
+
+
+```{r}
+filename <- mantis_report(df = example_prescription_numbers,
+                          file = "multipanel.html",
+                          inputspec = inputspec(
+                            timepoint_col = "PrescriptionDate",
+                            item_cols = c("Antibiotic", "Location"),
+                            value_col = "NumberOfPrescriptions"
+                          ),
+                          outputspec = outputspec_static_multipanel(
+                          ),
+                          report_title = "mantis report",
+                          dataset_description = "Antibiotic prescriptions by site",
+                          add_timestamp = TRUE,
+                          show_progress = TRUE
+)
+
+```
+
