### 2025-07-24 a8d4a60
Set knit_root_dir to subdirectory of tempdir too, though cannot see that this actually changes anything. (#24)
