### 2025-07-24 3fdf063
temporarily check what ubuntu filename looks like
