### 2025-07-29 e00a75e
update news
