### 2025-07-29 0cb8c77
Align to initial light-cran release (v0.4.2). Fixes #24 and #25
