+++ vignettes/articles/example-report-static-multipanel.Rmd
@@ -24,6 +24,7 @@ if (
   filename <-
     mantis::mantis_report(
       df = example_prescription_numbers,
+      file = "../../docs/articles/example-prescriptions-static-multipanel.html",
       inputspec = mantis::inputspec(
         timepoint_col = "PrescriptionDate",
         item_cols = c("Antibiotic", "Location"),
@@ -33,8 +34,6 @@ if (
       outputspec = mantis::outputspec_static_multipanel(),
       report_title = "Example mantis report - multipanel",
       dataset_description = "Antibiotic prescriptions by site",
-      save_directory = "../../docs/articles",
-      save_filename = "example-prescriptions-static-multipanel",
       show_progress = FALSE
     )
 }
