+++ README.Rmd
@@ -56,6 +56,10 @@ to capture programmatically.
 ## Installation
 
 ```{r, eval = FALSE}
+# install from CRAN
+install.packages("mantis")
+
+# or install the latest github release
 # install dependencies first
 install.packages(c(
   "remotes",
@@ -71,12 +75,10 @@ install.packages(c(
   "purrr",
   "htmltools"
 ))
-
-# install the latest release version
 remotes::install_github("phuongquan/mantis@v0.3.0", dependencies = FALSE)
 
 # or install the current development version
-# remotes::install_github("phuongquan/mantis", dependencies = FALSE)
+remotes::install_github("phuongquan/mantis", dependencies = FALSE)
 ```
 
 ## Usage
@@ -92,9 +94,10 @@ head(example_prescription_numbers)
 ```
 
 ```{r, eval = FALSE}
-# create a report in the current directory, with one tab per Location
+# create a report in the working directory, with one tab per Location
 mantis_report(
   df = example_prescription_numbers,
+  file = "example_prescription_numbers_report.html",
   inputspec = inputspec(
     timepoint_col = "PrescriptionDate",
     item_cols = c("Location", "Antibiotic", "Spectrum"),
