+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: ~
+url: https://phuongquan.github.io/mantis
 template:
   bootstrap: 5
 development:
