### 2025-07-31 3c5d3a9
Include bespoke reports functionality
