+++ R/construct.R
@@ -1,4 +1,283 @@
 # =============================================================================
+# EXPORTED FUNCTIONS
+
+# -----------------------------------------------------------------------------
+#' Initialise HTML widgets
+#'
+#' Since the output is being constructed in `results='asis'` chunks, there must
+#' also be at least one standard chunk that contains the relevant widgets,
+#' otherwise they will fail to render. The `dygraph` also needs to be
+#' initialised with the appropriate `plot_type`. This is only needed when
+#' creating interactive reports. Make sure you read the vignette:
+#' `vignette("bespoke-reports", package = "mantis")` as it contains further
+#' important information. Note: The chunk currently appears like a line break
+#' when rendered. See <https://github.com/rstudio/rmarkdown/issues/1877> for
+#' more info.
+#'
+#' @param plot_type "`bar`" or "`line`", depending on what will be used in real
+#'   tables. Or "none" if just want a reactable widget without dygraphs e.g. for
+#'   alerts
+#' @return A (mostly) invisible html widget
+#'
+#' @examples
+#' \dontrun{
+#' # put this inside its own chunk in the rmd file
+#' # it ensures that the dygraphs render when built using `cat()`
+#' # set the plot_type to the same plot_type as the real output
+#' mantis::bespoke_rmd_initialise_widgets(plot_type = "bar")
+#' }
+#'
+#' @export
+bespoke_rmd_initialise_widgets <- function(
+    plot_type
+) {
+  validate_params_required(match.call())
+  validate_params_type(
+    match.call(),
+    plot_type = plot_type
+  )
+
+  initialise_widgets(plot_type = plot_type)
+}
+
+
+# -----------------------------------------------------------------------------
+#' Dynamically generate mantis output for an rmd chunk
+#'
+#' Add `mantis` tabs and visualisations to an existing `rmarkdown` report. The
+#' function writes directly to the chunk using side-effects, so chunk options
+#' must contain `results = 'asis'`. Make sure you read the vignette:
+#' `vignette("bespoke-reports", package = "mantis")` as it contains further
+#' important information.
+#'
+#' @param df A data frame containing multiple time series in long format. See
+#'   Details.
+#' @param inputspec [`inputspec()`] object specifying which columns in the
+#'   supplied `df` represent the "timepoint", "item", and "value" for the time
+#'   series. A separate tab will be created for each distinct value in the "tab"
+#'   column.
+#' @param outputspec `outputspec` object specifying the desired format of the
+#'   html table(s)/plot(s). If not supplied, default values will be used.
+#' @param alertspec [`alertspec()`] object specifying conditions to test and
+#'   display
+#' @param timepoint_limits Set start and end dates for time period to include.
+#'   Defaults to min/max of `timepoint_col`
+#' @param fill_with_zero Logical. Replace any missing or `NA` values with 0?
+#'   Useful when `value_col` is a record count
+#' @param tab_name Character string to appear on the tab label. If omitted or
+#'   `NULL`, only the content/child tabs (and not the parent tab) will be
+#'   created.
+#' @param tab_level integer specifying the nesting level of the tab. If
+#'   `tab_name` is specified, a value of 1 generates a tab at rmd level "##",
+#'   and any `tab_col` tabs at a level down. If `tab_name` is not specified, any
+#'   `tab_col` tabs will be created at rmd level "##".
+#' @return (invisibly) the supplied `df`
+#'
[…]
