+++ NAMESPACE
@@ -9,6 +9,9 @@ export(alert_equals)
 export(alert_missing)
 export(alert_rules)
 export(alertspec)
+export(bespoke_rmd_alert_results)
+export(bespoke_rmd_initialise_widgets)
+export(bespoke_rmd_output)
 export(inputspec)
 export(mantis_alerts)
 export(mantis_report)

[…]
