### 2025-08-01 d0a69ef
align with CRAN release v0.4.3
