+++ dev/pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/example-report-static-heatmap: example-report-static-heatmap.html
   articles/example-report-static-multipanel: example-report-static-multipanel.html
   mantis: mantis.html
-last_built: 2025-07-29T12:27Z
+last_built: 2025-08-01T13:11Z
 urls:
   reference: https://phuongquan.github.io/mantis/reference
   article: https://phuongquan.github.io/mantis/articles
