### 2025-08-06 #710 @phuongquan
Hi @bozaah thanks very much for your review!

@beatrizmilz should I proceed with addressing these comments or should I wait until  a second review has been completed and do them all together?
