### 2025-08-08 1983413
only generate the summary_cols requested
