### 2025-08-08 ed8998d
deal with some minor todos
