+++ tests/testthat/test-prepare.R
@@ -875,49 +875,89 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
 })
 
 
-test_that("validate_df_to_inputspec() checks that item column can contain NA values", {
+test_that("validate_df_to_inputspec() checks that item column(s) can contain NA values", {
   df <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
     item = c(NA, NA, NA, rep(1, 7)),
+    item2 = c(NA, NA, NA, rep(1, 7)),
     value = rep(3, 10),
     stringsAsFactors = FALSE
   )
 
-  inputspec <- inputspec(
-    timepoint_col = "timepoint",
-    item_cols = "item",
-    value_col = "value"
-  )
-
+  # single item_col
   expect_silent(
     validate_df_to_inputspec(
       df = df,
-      inputspec = inputspec
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = "item",
+        value_col = "value"
+      )
     )
   )
+
+  # multi item_cols
+  expect_silent(
+    validate_df_to_inputspec(
+      df = df,
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = c("item", "item2"),
+        value_col = "value"
+      )
+    )
+  )
+
 })
 
-test_that("validate_df_to_inputspec() checks that item column doesn't contain both NA values and 'NA' strings", {
+test_that("validate_df_to_inputspec() checks that item column(s) don't contain both NA values and 'NA' strings", {
   df <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
     item = c(NA, NA, "NA", rep(1, 7)),
+    item2 = c(NA, NA, NA, rep(1, 7)),
+    item3 = c("NA", "NA", "NA", rep(1, 7)),
     value = rep(3, 10),
     stringsAsFactors = FALSE
   )
 
-  inputspec <- inputspec(
-    timepoint_col = "timepoint",
-    item_cols = "item",
-    value_col = "value"
-  )
-
+  # single item_col with NA and "NA"
   expect_error(
     validate_df_to_inputspec(
       df = df,
-      inputspec = inputspec
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = "item",
+        value_col = "value"
+      )
     ),
     class = "invalid_data"
   )
+
+  # multi item_cols with NA and "NA"
+  expect_error(
+    validate_df_to_inputspec(
+      df = df,
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = c("item", "item2"),
+        value_col = "value"
+      )
+    ),
+    class = "invalid_data"
+  )
+
+  # multi item_cols with NA and "NA" exclusively in separate cols
+  expect_silent(
+    validate_df_to_inputspec(
+      df = df,
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = c("item3", "item2"),
+        value_col = "value"
+      )
+    )
+  )
+
 })
 
 
