### 2025-08-12 7b3f640
finish off and add some tests
