+++ tests/testthat/test-prepare.R
@@ -1250,24 +1250,297 @@ test_that("arrange_items() sorts by two item_orders", {
 })
 
 
-test_that("adjust_timepoint_limit() moves supplied limit appropriately", {
+test_that("adjust_timepoint_limit() moves supplied limit appropriately for months", {
+  # min_timepoint before first timepoint
   expect_equal(
     adjust_timepoint_limit(
       timepoint_limit = as.Date("2022-01-01"),
-      timepoint_value = as.Date("2022-01-20"),
+      timepoint_values = seq(as.Date("2022-01-20"), by = "month", length.out = 5),
       timepoint_unit = "month",
-      direction = "earlier"
+      limit_type = "min"
+    ),
+    as.Date("2022-01-20")
+  )
+
+  # min_timepoint at first timepoint
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2022-01-20"),
+      timepoint_values = seq(as.Date("2022-01-20"), by = "month", length.out = 5),
+      timepoint_unit = "month",
+      limit_type = "min"
+    ),
+    as.Date("2022-01-20")
+  )
+
+  # min_timepoint between timepoint_values with gaps
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2022-05-01"),
+      timepoint_values = c(as.Date("2022-01-20"), as.Date("2023-01-20")),
+      timepoint_unit = "month",
+      limit_type = "min"
+    ),
+    as.Date("2022-05-20")
+  )
+
+  # min_timepoint after last timepoint
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2023-01-01"),
+      timepoint_values = seq(as.Date("2022-01-20"), by = "month", length.out = 5),
+      timepoint_unit = "month",
+      limit_type = "min"
+    ),
+    as.Date("2023-01-20")
+  )
+
+  # max_timepoint after last timepoint
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2023-01-01"),
+      timepoint_values = seq(as.Date("2022-01-20"), by = "month", length.out = 5),
+      timepoint_unit = "month",
+      limit_type = "max"
+    ),
+    as.Date("2022-12-20")
+  )
+
+  # max_timepoint between timepoint_values with gaps
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2022-05-01"),
+      timepoint_values = c(as.Date("2022-01-20"), as.Date("2023-01-20")),
+      timepoint_unit = "month",
+      limit_type = "max"
+    ),
+    as.Date("2022-04-20")
+  )
+
+  # max_timepoint before first timepoint
+  expect_equal(
+    adjust_timepoint_limit(
+      timepoint_limit = as.Date("2022-01-01"),
+      timepoint_values = seq(as.Date("2022-01-20"), by = "month", length.out = 5),
+      timepoint_unit = "month",
+      limit_type = "max"
     ),
     as.Date("2021-12-20")
   )
+
+})
+
+test_that("adjust_timepoint_limit() moves supplied limit appropriately for weeks", {
+  # min_timepoint before first timepoint
   expect_equal(
     adjust_timepoint_limit(
       timepoint_limit = as.Date("2022-01-01"),
-      timepoint_value = as.Date("2022-01-20"),
-      timepoint_unit = "month",
-      direction = "later"
+      timepoint_values = seq(as.Date("2022-01-20"), by = "week", length.out = 5),
[…]
