### 2025-08-13 7735bf5
check for mixture of date/posixt classes
