### 2025-08-13 dae6409
Adjust timepoint limits (#27)
