+++ R/prepare.R
@@ -699,28 +699,35 @@ align_data_timepoints <- function(
   # initialise column names to avoid R CMD check Notes
   timepoint <- value <- NULL
 
-  # TODO: Need to work out correct limits to use based on df
-  #  in case supplied limits don't match df granularity
-  if (is.na(timepoint_limits[1])) {
-    min_timepoint <- min(prepared_df$timepoint)
-  } else {
-    min_timepoint <- timepoint_limits[1]
+  min_timepoint <- min(prepared_df$timepoint)
+  max_timepoint <- max(prepared_df$timepoint)
+  # if the supplied limit(s) don't match the df granularity, be kind and adjust it
+  if (!is.na(timepoint_limits[1])) {
+    min_timepoint <- adjust_timepoint_limit(
+      timepoint_limit = timepoint_limits[1],
+      timepoint_values = prepared_df$timepoint,
+      timepoint_unit = inputspec$timepoint_unit,
+      limit_type = "min")
   }
-  if (is.na(timepoint_limits[2])) {
-    max_timepoint <- max(prepared_df$timepoint)
-  } else {
+  if (!is.na(timepoint_limits[2])) {
     # NOTE: While timepoint_limits should already be a date class,
     # if user supplies an NA first in the vector, the second value gets coerced
     # to numeric and leads to an error in seq() later on
     if (inputspec$timepoint_unit %in% c("sec", "min", "hour")) {
-      max_timepoint <- as.POSIXct(timepoint_limits[2])
+      max_timepoint <- adjust_timepoint_limit(
+        timepoint_limit = as.POSIXct(timepoint_limits[2]),
+        timepoint_values = prepared_df$timepoint,
+        timepoint_unit = inputspec$timepoint_unit,
+        limit_type = "max")
     } else {
-      max_timepoint <- as.Date(timepoint_limits[2])
+      max_timepoint <- adjust_timepoint_limit(
+        timepoint_limit = as.Date(timepoint_limits[2]),
+        timepoint_values = prepared_df$timepoint,
+        timepoint_unit = inputspec$timepoint_unit,
+        limit_type = "max")
     }
   }
 
-  # TODO: Need to work out correct granularity to use based on df
-  #  as don't want to insert unnecessary rows
   all_timepoints <- seq(
     min_timepoint,
     max_timepoint,
@@ -759,6 +766,83 @@ align_data_timepoints <- function(
     )
 
   df_out
+}
+
+# -----------------------------------------------------------------------------
+#' Adjust the supplied timepoint_limit to align with data if necessary
+#'
+#' @param timepoint_limit datetime value supplied by user
+#' @param timepoint_values datetime values in the data
+#' @param timepoint_unit granularity specified by user
+#' @param limit_type "min" or "max"
+#'
+#' @returns Datetime
+#' @noRd
+adjust_timepoint_limit <- function(
+  timepoint_limit,
+  timepoint_values,
+  timepoint_unit,
+  limit_type
+) {
+
+  # if everything is in days then no changes needed
+  if (
+    timepoint_unit == "day" &&
+      inherits(timepoint_limit, what = "Date") && inherits(timepoint_values, what = "Date")
+  ) {
+    return(timepoint_limit)
+  }
+
+  # ensure limit is of same class as values
+  if (inherits(timepoint_values, what = "POSIXct")) {
+    timepoint_limit <- as.POSIXct(timepoint_limit)
+  } else if (inherits(timepoint_values, what = "POSIXlt")) {
+    timepoint_limit <- as.POSIXlt(timepoint_limit)
+  } else if (inherits(timepoint_values, what = "Date")) {
+    timepoint_limit <- as.Date(timepoint_limit)
+  }
+
[…]
