+++ README.Rmd
@@ -75,7 +75,7 @@ install.packages(c(
   "purrr",
   "htmltools"
 ))
-remotes::install_github("phuongquan/mantis@v0.3.0", dependencies = FALSE)
+remotes::install_github("phuongquan/mantis@v0.4.3", dependencies = FALSE)
 
 # or install the current development version
 remotes::install_github("phuongquan/mantis", dependencies = FALSE)
