+++ README.md
@@ -57,6 +57,10 @@ We anticipate two main target users:
 ## Installation
 
 ``` r
+# install from CRAN
+install.packages("mantis")
+
+# or install the latest github release
 # install dependencies first
 install.packages(c(
   "remotes",
@@ -72,12 +76,10 @@ install.packages(c(
   "purrr",
   "htmltools"
 ))
-
-# install the latest release version
-remotes::install_github("phuongquan/mantis@v0.3.0", dependencies = FALSE)
+remotes::install_github("phuongquan/mantis@v0.4.3", dependencies = FALSE)
 
 # or install the current development version
-# remotes::install_github("phuongquan/mantis", dependencies = FALSE)
+remotes::install_github("phuongquan/mantis", dependencies = FALSE)
 ```
 
 ## Usage
@@ -103,9 +105,10 @@ head(example_prescription_numbers)
     ## 6 2022-01-01       Vancomycin    Limited                      0 SITE1
 
 ``` r
-# create a report in the current directory, with one tab per Location
+# create a report in the working directory, with one tab per Location
 mantis_report(
   df = example_prescription_numbers,
+  file = "example_prescription_numbers_report.html",
   inputspec = inputspec(
     timepoint_col = "PrescriptionDate",
     item_cols = c("Location", "Antibiotic", "Spectrum"),
