### 2025-08-15 #710 @phuongquan
Hi @beatrizmilz, as I mentioned previously, due to local time pressures, I have had to proceed with submitting mantis to CRAN, and it has now been accepted there. The CRAN submission process required me to make two breaking changes compared to the version that I initially submitted here (0.3.0), and is the cause of the two errors that @bozaah found in the readme and website. I have now updated the main branch and the website to align with the CRAN release (0.4.3) so that everything matches again. Can the next reviewer be asked to review v0.4.3 instead of v0.3.0, as I think this will make things less confusing?
