+++ dev/pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/example-report-static-heatmap: example-report-static-heatmap.html
   articles/example-report-static-multipanel: example-report-static-multipanel.html
   mantis: mantis.html
-last_built: 2025-08-13T16:36Z
+last_built: 2025-08-15T10:38Z
 urls:
   reference: https://phuongquan.github.io/mantis/reference
   article: https://phuongquan.github.io/mantis/articles
