### 2025-09-17 2a254fe
minor rendering differences in ggplot2 v4
