### 2025-09-17 95ee748
update arrange_items() to work with >3 items
