### 2025-09-19 bdde915
handle Posixt values in max_else_na()
