### 2025-09-19 8335c02
Be stricter about timezones. Fixes #30
