+++ tests/testthat/test-prepare.R
@@ -309,8 +309,8 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # with time portion (POSIXct)
   df <- data.frame(
     timepoint = seq(
-      as.POSIXct("2022-01-01 12:00:00"),
-      as.POSIXct("2022-01-10 12:00:00"),
+      as.POSIXct("2022-01-01 12:00:00", tz = "UTC"),
+      as.POSIXct("2022-01-10 12:00:00", tz = "UTC"),
       by = "days"
     ),
     item = rep(1, 10),
@@ -328,8 +328,8 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # with time portion (POSIXlt)
   df <- data.frame(
     timepoint = seq(
-      as.POSIXlt("2022-01-01 12:00:00"),
-      as.POSIXlt("2022-01-10 12:00:00"),
+      as.POSIXlt("2022-01-01 12:00:00", tz = "UTC"),
+      as.POSIXlt("2022-01-10 12:00:00", tz = "UTC"),
       by = "days"
     ),
     item = rep(1, 10),
@@ -348,9 +348,9 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # different time each day
   df <- data.frame(
     timepoint = c(
-      as.POSIXlt("2022-01-01 12:00:00"),
-      as.POSIXlt("2022-01-02 13:00:00"),
-      as.POSIXlt("2022-01-10 16:00:00")
+      as.POSIXlt("2022-01-01 12:00:00", tz = "UTC"),
+      as.POSIXlt("2022-01-02 13:00:00", tz = "UTC"),
+      as.POSIXlt("2022-01-10 16:00:00", tz = "UTC")
     ),
     item = rep(1, 3),
     value = rep(3, 3),
@@ -364,6 +364,7 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
     ),
     class = "invalid_data"
   )
+
 })
 
 
@@ -412,8 +413,8 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # with time portion
   df <- data.frame(
     timepoint = seq(
-      as.POSIXct("2022-01-01 12:00:00"),
-      as.POSIXct("2022-03-05 12:00:00"),
+      as.POSIXct("2022-01-01 12:00:00", tz = "UTC"),
+      as.POSIXct("2022-03-05 12:00:00", tz = "UTC"),
       by = "week"
     ),
     item = rep(1, 10),
@@ -432,9 +433,9 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # different time each day
   df <- data.frame(
     timepoint = c(
-      as.POSIXlt("2022-01-01 12:00:00"),
-      as.POSIXlt("2022-01-08 13:00:00"),
-      as.POSIXlt("2022-01-15 16:00:00")
+      as.POSIXlt("2022-01-01 12:00:00", tz = "UTC"),
+      as.POSIXlt("2022-01-08 13:00:00", tz = "UTC"),
+      as.POSIXlt("2022-01-15 16:00:00", tz = "UTC")
     ),
     item = rep(1, 3),
     value = rep(3, 3),
@@ -497,8 +498,8 @@ test_that("validate_df_to_inputspec() checks that timepoint_col matches inputspe
   # with time portion
   df <- data.frame(
     timepoint = seq(
-      as.POSIXct("2022-01-02 12:00:00"),
-      as.POSIXct("2022-10-02 12:00:00"),
+      as.POSIXct("2022-01-02 12:00:00", tz = "UTC"),
[…]
