### 2025-09-20 5ee82f6
Use value recycling for test dfs where possible
