+++ tests/testthat/test-prepare.R
@@ -1,8 +1,8 @@
 test_that("prepare_table() avoids min/max warnings when all values in a group are NA", {
   df <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
-    item_zero = rep(0, 10),
-    item_na = rep(NA, 10),
+    item_zero = 0,
+    item_na = NA,
     stringsAsFactors = FALSE
   ) |>
     tidyr::pivot_longer(
@@ -33,11 +33,11 @@ test_that("validate_df_to_inputspec() checks that duplicate column names in data
         as.Date("2022-01-10"),
         by = "days"
       ),
-      item = rep(1, 10),
-      value = rep(3, 10),
+      item = 1,
+      value = 3,
       stringsAsFactors = FALSE
     ),
-    data.frame(item = rep(2, 10), stringsAsFactors = FALSE)
+    data.frame(item = 2, stringsAsFactors = FALSE)
   )
 
   inputspec <- inputspec(
@@ -58,8 +58,8 @@ test_that("validate_df_to_inputspec() checks that duplicate column names in data
 test_that("validate_df_to_inputspec() checks that duplicate column names in inputspec not allowed", {
   df <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
-    item = rep(1, 10),
-    value = rep(3, 10),
+    item = 1,
+    value = 3,
     stringsAsFactors = FALSE
   )
 
@@ -95,8 +95,8 @@ test_that("validate_df_to_inputspec() checks that duplicate column names in inpu
 test_that("validate_df_to_inputspec() checks that supplied colnames are present in df", {
   df <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
-    item = rep(1, 10),
-    value = rep(3, 10),
+    item = 1,
+    value = 3,
     group = rep("a", 10),
     stringsAsFactors = FALSE
   )
@@ -149,8 +149,8 @@ test_that("validate_df_to_inputspec() checks that supplied colnames are present 
 test_that("validate_df_to_inputspec() checks that timepoint_col is a datetime type", {
   df_orig <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
-    item = rep(1, 10),
-    value = rep(3, 10),
+    item = 1,
+    value = 3,
     stringsAsFactors = FALSE
   )
   df <- df_orig
@@ -208,8 +208,8 @@ test_that("validate_df_to_inputspec() checks that timepoint_col is a datetime ty
 test_that("validate_df_to_inputspec() checks that value_col is a numeric type", {
   df_orig <- data.frame(
     timepoint = seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
-    item = rep(1, 10),
-    value = rep(3, 10),
+    item = 1,
+    value = 3,
     stringsAsFactors = FALSE
   )
   df <- df_orig
@@ -262,8 +262,8 @@ test_that("validate_df_to_inputspec() checks that timepoint column doesn't conta
       seq(as.Date("2022-01-01"), as.Date("2022-01-09"), by = "days"),
       NA
     ),
-    item = rep(1, 10),
-    value = rep(3, 10),
+    item = 1,
+    value = 3,
     stringsAsFactors = FALSE
   )
 
[…]
