### 2025-09-20 539aad0
refactor creation of alert_results df
