### 2025-09-20 9c3e491
in alerts, rename function_call to expression
