+++ R/alert.R
@@ -302,10 +302,10 @@ alert_missing <- function(
   )
 
   if (extent_type == "all") {
-    function_call <- quote(all(is.na(value)))
+    expression <- quote(all(is.na(value)))
     rule_description <- "All values are missing"
   } else if (extent_type == "any") {
-    function_call <- substitute(
+    expression <- substitute(
       sum(is.na(value)) >= x,
       list(x = extent_value)
     )
@@ -315,7 +315,7 @@ alert_missing <- function(
       " values are missing in total"
     )
   } else if (extent_type == "last") {
-    function_call <- substitute(
+    expression <- substitute(
       all(is.na(rev(value)[1:x])),
       list(x = extent_value)
     )
@@ -325,7 +325,7 @@ alert_missing <- function(
       " or more values are missing"
     )
   } else if (extent_type == "consecutive") {
-    function_call <- substitute(
+    expression <- substitute(
       {
         run_lengths <- rle(is.na(value))
         any(run_lengths$lengths[run_lengths$values] >= x)
@@ -340,7 +340,7 @@ alert_missing <- function(
 
   alert_rule(
     type = "missing",
-    function_call = function_call,
+    expression = expression,
     items = items,
     short_name = rule_short_name,
     description = rule_description
@@ -399,7 +399,7 @@ alert_equals <- function(
   )
 
   if (extent_type == "all") {
-    function_call <- substitute(
+    expression <- substitute(
       all(value[!is.na(value)] == rv),
       list(rv = rule_value)
     )
@@ -408,7 +408,7 @@ alert_equals <- function(
       rule_value
     )
   } else if (extent_type == "any") {
-    function_call <- substitute(
+    expression <- substitute(
       sum(value[!is.na(value)] == rv) >= x,
       list(x = extent_value, rv = rule_value)
     )
@@ -419,7 +419,7 @@ alert_equals <- function(
       rule_value
     )
   } else if (extent_type == "last") {
-    function_call <- substitute(
+    expression <- substitute(
       all(rev(value[!is.na(value)])[1:x] == rv),
       list(x = extent_value, rv = rule_value)
     )
@@ -430,7 +430,7 @@ alert_equals <- function(
       rule_value
     )
   } else if (extent_type == "consecutive") {
-    function_call <- substitute(
+    expression <- substitute(
       {
         run_lengths <- rle(value[!is.na(value)])
         any(run_lengths$lengths[run_lengths$values == rv] >= x)
@@ -446,7 +446,7 @@ alert_equals <- function(
 
   alert_rule(
     type = "equals",
-    function_call = function_call,
+    expression = expression,
     items = items,
     short_name = rule_short_name,
     description = rule_description
@@ -500,7 +500,7 @@ alert_above <- function(
   )
 
   if (extent_type == "all") {
-    function_call <- substitute(
+    expression <- substitute(
       all(value[!is.na(value)] > rv),
       list(rv = rule_value)
     )
@@ -509,7 +509,7 @@ alert_above <- function(
       rule_value
     )
   } else if (extent_type == "any") {
-    function_call <- substitute(
+    expression <- substitute(
       sum(value[!is.na(value)] > rv) >= x,
       list(x = extent_value, rv = rule_value)
     )
@@ -520,7 +520,7 @@ alert_above <- function(
       rule_value
     )
   } else if (extent_type == "last") {
-    function_call <- substitute(
+    expression <- substitute(
       all(rev(value[!is.na(value)])[1:x] > rv),
       list(x = extent_value, rv = rule_value)
     )
@@ -531,7 +531,7 @@ alert_above <- function(
       rule_value
     )
   } else if (extent_type == "consecutive") {
-    function_call <- substitute(
+    expression <- substitute(
[…]
