### 2025-09-20 06580c6
more NAs in alert rule tests
