### 2025-09-20 c7055c8
move remaining TODOs to issues
