### 2025-09-23 baaad1a
Update vignettes
