### 2025-09-23 4f03d4f
update news and description
