+++ NEWS.md
@@ -1,6 +1,17 @@
 # mantis (development version)
 
-Re-align with CRAN release
+Incorporates changes requested for acceptance into https://ropensci.org/.
+
+## Breaking changes
+
+* `expression` parameter of `alert_custom()` replaces `function_call`
+
+## Bug fixes and minor improvements
+
+* `item_order` parameter of `outputspec()` can now handle an arbitrary number of items
+* `timepoint_limits` parameter for alerts and bespoke reports now automatically adjusts user-supplied values to align with the data and `timepoint_unit` where necessary, instead of inserting inappropriate additional timepoints in plots
+* Date-only timepoints cast as POSIXt using a daylight-savings-varying timezone no longer result in inappropriate additional timepoints appearing in plots (#30)
+* Stricter validation of POSIXt timepoints
 
 # mantis 0.4.3 (2025-07-31)
 
