+++ DESCRIPTION
@@ -1,12 +1,14 @@
 Package: mantis
 Type: Package
 Title: Multiple Time Series Scanner
-Version: 0.5.0.9000
+Version: 0.5.0.9001
 Authors@R: c(
     person(c("T.", "Phuong"), "Quan", email = "phuong.quan@ndm.ox.ac.uk",
         role = c("aut", "cre"), comment = c(ORCID = "0000-0001-8566-1817")),
     person(family = "University of Oxford", role = "cph"),
-    person(family = "National Institute for Health Research (NIHR)", role = "fnd")
+    person(family = "National Institute for Health Research (NIHR)", role = "fnd"),
+    person("Rodrigo", "Pires", role = "rev"),
+    person("Margaret", "Siple", role = "rev")
     )
 Description: Generate interactive html reports that enable quick visual review of multiple related time series stored in a data frame. For static datasets, this can help to identify any temporal artefacts that may affect the validity of subsequent analyses. For live data feeds, regularly scheduled reports can help to pro-actively identify data feed problems or unexpected trends that may require action. The reports are self-contained and shareable without a web server. 
 URL: https://github.com/phuongquan/mantis,
@@ -36,6 +38,6 @@ Suggests:
     vdiffr,
     withr
 Config/testthat/edition: 3
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
