### 2025-10-03 #710 @beatrizmilz
> Hi [@mpadge](https://github.com/mpadge) [@beatrizmilz](https://github.com/beatrizmilz) I asked the review bot to submit my response yesterday but it doesn't appear to have acted on it. Do I need to do something different?

Hi @phuongquan ! I'm late with this response. I'll write back soon. I'm sorry for the delay.
