### 2025-10-04 #710 @beatrizmilz
Thank you @phuongquan for [responding](https://github.com/ropensci/software-review/issues/710#issuecomment-3324852026) to the reviewers’ feedback!

Hi @bozaah and @MargaretSiple-NOAA, could you please indicate whether you approve the package, using the reviewer approval template (pasted below)?

Thank you!

```
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing:
```
