+++ _pkgdown.yml
@@ -64,6 +64,6 @@ reference:
 
 authors:
   footer:
-    roles: [cre, cph]
+    roles: [cre]
   sidebar:
-    roles: [cre, cph]
+    roles: [cre]
