+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Multiple Time Series Scanner
 Version: 0.5.0.9001
 Authors@R: c(
-    person(c("T.", "Phuong"), "Quan", email = "phuong.quan@ndm.ox.ac.uk",
+    person(c("T.", "Phuong"), "Quan", email = "phuongquan567@outlook.com",
         role = c("aut", "cre"), comment = c(ORCID = "0000-0001-8566-1817")),
     person(family = "University of Oxford", role = "cph"),
     person(family = "National Institute for Health Research (NIHR)", role = "fnd"),
