+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mantis
 Type: Package
 Title: Multiple Time Series Scanner
-Version: 0.5.0.9001
+Version: 1.0.0
 Authors@R: c(
     person(c("T.", "Phuong"), "Quan", email = "phuongquan567@outlook.com",
         role = c("aut", "cre"), comment = c(ORCID = "0000-0001-8566-1817")),
@@ -35,7 +35,10 @@ Imports:
 Suggests: 
     covr,
     testthat (>= 3.0.0),
+    codemetar,
     vdiffr,
+    shiny,
+    diffviewer,
     withr
 Config/testthat/edition: 3
 RoxygenNote: 7.3.3
