### 2025-10-27 #710 @phuongquan
Hi @beatrizmilz @bozaah @MargaretSiple-NOAA,

Due to local time pressures I have needed to submit the current version of mantis (as described in my response above) to CRAN. I am still happy to make any further changes required though. Thanks.
