### 2025-11-09 #710 @beatrizmilz
> Hi [@bozaah](https://github.com/bozaah) and [@MargaretSiple-NOAA](https://github.com/MargaretSiple-NOAA)! Just checking in to see if you’ve had a chance to take a look at the final updates. When you have a moment, could you please let us know whether you approve the package using the [reviewer approval template above](https://github.com/ropensci/software-review/issues/710#issuecomment-3368586415)?
>
> Thanks again for all your time and thoughtful reviews, we really appreciate it!

Hi @MargaretSiple-NOAA !
Just following up to check whether you’ve had a chance to read this message.
Thank you!
