### 2025-11-10 #710 @mpadge
@beatrizmilz Unfortunately, @MargaretSiple-NOAA is out of action due to US govt shutdown, but she will hopefully be able to respond from private account at @mcsiple
