### 2025-12-04 c9e6bd0
Fix error when timepoint_limits[1] is NA. Fix #35
