+++ dev/reference/mantis_report.md
@@ -1,0 +1,209 @@
+# Create an interactive time series report from a data frame
+
+Accepts a data frame containing multiple time series in long format,
+generates a collection of interactive time series plots for visual
+inspection, and saves the report to disk.
+
+## Usage
+
+``` r
+mantis_report(
+  df,
+  file,
+  inputspec,
+  outputspec = NULL,
+  alertspec = NULL,
+  report_title = "mantis report",
+  dataset_description = "",
+  add_timestamp = FALSE,
+  show_progress = TRUE,
+  ...
+)
+```
+
+## Arguments
+
+- df:
+
+  A data frame containing multiple time series in long format. See
+  Details.
+
+- file:
+
+  String specifying the desired file name (and path) to save the report
+  to. The file name should include the extension ".html". If only a file
+  name is supplied, the report will be saved in the current working
+  directory. If a path is supplied, the directory should already exist.
+  Any existing file of the same name will be overwritten unless
+  `add_timestamp` is set to `TRUE`.
+
+- inputspec:
+
+  [`inputspec()`](https://phuongquan.github.io/mantis/dev/reference/inputspec.md)
+  object specifying which columns in the supplied `df` represent the
+  "timepoint", "item", "value" and (optionally) "tab" for the time
+  series. If a "tab" column is specified, a separate tab will be created
+  for each distinct value in the column.
+
+- outputspec:
+
+  `outputspec` object specifying the desired format of the html
+  table(s). If not supplied, default values will be used.
+
+- alertspec:
+
+  [`alertspec()`](https://phuongquan.github.io/mantis/dev/reference/alertspec.md)
+  object specifying conditions to test and display.
+
+- report_title:
+
+  Title to appear on the report.
+
+- dataset_description:
+
+  Short description of the dataset being shown. This will appear on the
+  report.
+
+- add_timestamp:
+
+  Append a timestamp to the end of the filename with format
+  `_YYMMDD_HHMMSS`. This can be used to keep multiple versions of the
+  same report. Default = `FALSE`.
+
+- show_progress:
+
+  Print progress to console. Default = `TRUE`.
+
+- ...:
+
+  Further parameters to be passed to
+  [`rmarkdown::render()`](https://pkgs.rstudio.com/rmarkdown/reference/render.html).
+  Cannot include any of `input`, `output_dir`, `output_file`, `params`,
+  `quiet`.
+
+## Value
+
+A string containing the name and full path of the saved report.
+
+## Details
+
+The supplied data frame should contain multiple time series in long
+format, i.e.:
+
+- one "timepoint" (date/posixt) column which will be used for the
+  x-axes. Values should follow a regular pattern, e.g. daily or monthly,
+  but do not have to be consecutive.
+
+- one or more "item" (character) columns containing categorical values
+  identifying distinct time series.
+
+- one "value" (numeric) column containing the time series values which
+  will be used for the y-axes.
+
+- Optionally, a "tab" (character) column containing categorical values
+  which will be used to group the time series into different tabs on the
+  report.
+
+  The `inputspec` parameter maps the data frame columns to the above.
+
+## See also
+
+[`inputspec()`](https://phuongquan.github.io/mantis/dev/reference/inputspec.md),
[…]
