### 2025-12-04 #710 @phuongquan
Hi @MargaretSiple-NOAA ,

I believe I have identified (and fixed) the causes of the errors you found:

> I found a couple of other things while testing the bespoke reports vignette that I want to get clarification on before I approve:
> * Cloning `mantis-templates` and knitting basic-template.Rmd quits from L 57-84

Did you try to knit the Rmd directly from within e.g. RStudio? This file should actually have been rendered by running `main.R`. I have clarified this in the Readme and added a second file `basic-template-standalone.Rmd` that can be knitted directly.

> * Embedding as-is R chunk from the bespoke-reports vignette (first chunk after "[Adding `mantis` content](https://phuongquan.github.io/mantis/articles/bespoke-reports.html#adding-mantis-content)") into an Rmd `{r, results='asis'}` chunk returns an error. Adding the two required chunks to that same Rmd did not change the outcome

I believe I introduced a bug in my last set of changes relating to timezones. I have updated the code to fix this. Please install the latest version of mantis from github and try again. If you still get an error, please let me know exactly what the error says.

Thanks!
Phuong
