### 2025-12-14 #710 @ropensci-review-bot
Approved! Thanks @phuongquan for submitting and @bozaah, @MargaretSiple-NOAA for your reviews! :grin:
