### 2025-12-15 #710 @phuongquan
Hi @beatrizmilz, I have started on the to-dos and wanted to check if I should also be transferring my related repository https://github.com/phuongquan/mantis-templates to the ropensci organisation too? The `mantis-templates` repository is a helper repo that contains template markdown files to help `mantis` users to create their own bespoke reports (using `mantis` functions), and is linked to in one of the vignettes (https://ropensci.github.io/mantis/articles/bespoke-reports.html). Thanks.
