### 2026-01-20 #710 @mpadge
@phuongquan Congratulations on getting your package through! @beatrizmilz just asked for my opinion on the `mantis-templates` repo, so here it is:

I think that the contents of that should go within your main package. Putting within an `inst/extdata` folder should suffice, or `inst/extdata/templates` if you like. You can then write a function within the `mantis` package itself which grabs this data and copies to some specified local directory. See https://r-pkgs.org/data.html#sec-data-extdata for details, and subsequent section https://r-pkgs.org/data.html#sec-data-system-file for how to access those from within your package once it's installed. I'd suggest opening an issue in https://github.com/ropensci/mantis to tackle this, adding the URL of this issue so it'll be visible here, and pinging either @beatrizmilz or myself from there if you need any help. Thanks!
